refactor: add base config (#72)
for configs that are easy to override

We've seen CI fail for cloud-functions because postUpdateOptions
contains 'yarnDedupeFewer' in the default renovate config, but for some
reason doing that lockfile deduplication causes breaking changes. (I've
been unable to repro the test failures locally, but it happens
consistently during CI). Unfortunately, it is [not possible to
override](https://github.com/renovatebot/renovate/discussions/25112#discussioncomment-7239388)
`postUpdateOptions`. As a workaround, I created a "base" config that
includes only easy-to-override configs, and a "default" config that
extends it. This is a non-breaking change for all existing repos that
use valora's default config, and for cloud-functions we can extend
`base.json5` instead, which will let us omit `postUpdateOptions` for
now.